+++ src/adapters/controllers/theaterController.ts
@@ -1,8 +1,8 @@
 import { Request, Response } from "express";
 import { ITheaterAuth, ITheaterUpdate } from "../../interfaces/schema/theaterSchema";
 import { TheaterUseCase } from "../../useCases/theaterUseCase";
-import { STATUS_CODES } from "../../constants/httpStausCodes";
-import { TheaterShowLimit, maxDistance } from "../../constants/constants";
+import { STATUS_CODES } from "../../infrastructure/constants/httpStausCodes";
+import { TheaterShowLimit, maxDistance } from "../../infrastructure/constants/constants";
 import { ITempTheaterReq } from "../../interfaces/schema/tempTheaterSchema";
 import { ICoords, ITheaterAddress } from "../../entities/common";
