+++ src/infrastructure/db/movieModel.ts
@@ -1,6 +1,6 @@
 import mongoose, { Model, Schema } from "mongoose";
 import { IMovie } from "../../interfaces/schema/movieSchema";
-import { Languages } from "../../constants/langAbbreviation";
+import { Languages } from "../constants/langAbbreviation";
 import { reviewSchema } from "./subSchema/reviewSchema";
