+++ src/infrastructure/errors/cancelledByUnknownError.ts
@@ -1,4 +1,4 @@
-import { STATUS_CODES } from "../../constants/httpStausCodes";
+import { STATUS_CODES } from "../constants/httpStausCodes";
 import { CustomError } from "./customError";
 
 export class CancelledByUnknownError extends CustomError {
