+++ src/useCases/userUseCase.ts
@@ -1,7 +1,7 @@
 // import { AuthRes } from "../Types/AuthRes";
 import { log } from "console";
-import { OTP_TIMER } from "../constants/constants";
-import { STATUS_CODES } from "../constants/httpStausCodes";
+import { OTP_TIMER } from "../infrastructure/constants/constants";
+import { STATUS_CODES } from "../infrastructure/constants/httpStausCodes";
 import { get200Response, get500Response, getErrorResponse } from "../infrastructure/helperFunctions/response";
 import { TempUserRepository } from "../infrastructure/repositories/tempUserRepository";
 import { UserRepository } from "../infrastructure/repositories/userRepository";
